# docs/swagger.yaml
# this is an example of the Uber API
# as a demonstration of an API spec in YAML
swagger: '2.0'
info:
  title: Rating Service
  description: Rating service
  version: "0.0.1"
# the domain of the service
host: transafe.rafaelferreira.pt
# array of all schemes that your API supports
schemes:
  - http
# will be prefixed to all paths
paths:
  /rating:
    post:
      summary: Rate User
      description: |
        Method used to rate a user
      parameters:
        - name: dest_id
          description: Id of the rating target
          required: true
          type: string
          in: formData
        - name: source_id
          description: Id of the rating provider
          required: true
          type: string
          in: formData
        - name: rating
          description: Numeric rating value given (1-5)
          required: true
          type: number
          in: formData
        - name: message
          description: Additional message the source wants to provide. Max xx characters.
          required: false
          type: string
          in: formData
      responses:
        200:
          description: Rating successfully added.
        400:
          description: Invalid parameters.
          
  /rating/{id}:
    get:
      summary: Rate User
      description: |
        Method used to retrieve ratings
      produces:
        - application/json
      parameters:
        - name: id
          description: Id of the rating target
          required: true
          type: string
          in: path
        - name: rating
          description: Choice if user wants rating given, received or all. Default is received only.
          required: false
          type: string
          in: query
        - name: size
          description: How many results the user wants to receive or all. Only aplicable if results are more than the number specified. Default is latest 5 ratings. Number provided is always counting from latests.
          required: false
          type: string
          in: query
        - name: fields
          description: The fields the user wants to retrieve. Options are rating, source_id, dest_id, message. Default only rating is given
          required: false
          type: string
          in: query

      responses:
        200:
          description: Json data requested by the user
          schema: 
            $ref: "#/definitions/Rating"
        400:
          description: Invalid parameters.
      

definitions:
  Rating:
    properties:
      rating:
        type: number
        description: Rating number between 1-5.
      source_id:
        type: string
        description: ID of the user that gave rated.
      dest_id:
        type: string
        description: ID of the user that was rated.
      message:
        type: string
        description: Optional message provided by source_id about the rate.
      
        

    